Oops. Use the proper _FILE_OFFSET_BITS define name, and remove the thus useless explicit cast.

+++ KindleTool/create.c
@@ -240,7 +240,7 @@ int kindle_create_package_archive(const char *outname, char **filename, const in
                 // Don't add the bundlefile to itself
                 if(i != total_files - 1)
                 {
-                    if(fprintf(bundlefile, "%d %s %s %lld %s\n", ((IS_SCRIPT(pathname) || IS_SHELL(pathname)) ? 129 : 128), md5, pathname, (long long int)st.st_size / BLOCK_SIZE, sourcepath) < 0) // FIXME: The python script does pathname+"_file" for the last field.
+                    if(fprintf(bundlefile, "%d %s %s %lld %s\n", ((IS_SCRIPT(pathname) || IS_SHELL(pathname)) ? 129 : 128), md5, pathname, st.st_size / BLOCK_SIZE, sourcepath) < 0) // FIXME: The python script does pathname+"_file" for the last field.
                     {
                         fprintf(stderr, "Cannot write to index file.\n");
                         return -1;  // FIXME
